# Implement performance metrics collector workflow