--- a/hack/e2e/metrics/storageclass.yaml
+++ b/hack/e2e/metrics/storageclass.yaml
@@ -1,0 +1,6 @@
+apiVersion: storage.k8s.io/v1
+kind: StorageClass
+metadata:
+  name: ebs-sc
+provisioner: ebs.csi.aws.com
+volumeBindingMode: Immediate