**What is this PR about? / Why do we need it?**

This PR implements a workflow to collect driver deployment time, container resource usage, and volume lifecycle performance during CI runs with the goal of preventing silent regressions over time. At a high level, the metrics collector script:

- Collects driver deployment time, resource usage metrics, and clusterloader2 metrics.
- Configures bucket policy to allow the associated Lambda function to access the bucket.
- Sets a lifecycle policy on the S3 bucket to automatically expire metrics data after 3 days.
- Configures bucket notification to trigger the Lambda function on object creation events.
- Uploads the collected metrics to the S3 bucket.

**What testing is done?** 

```
I0612 14:54:08.101590   27309 phase_latency.go:146] PodStartupLatency: perc50: 11.72959236s, perc90: 14.944266027s, perc99: 14.944266027s; threshold 5m0s
I0612 14:54:08.101598   27309 phase_latency.go:146] PodStartupLatency: perc50: 4s, perc90: 4s, perc99: 4s
I0612 14:54:18.302610   27309 clusterloader.go:248] --------------------------------------------------------------------------------
I0612 14:54:18.302627   27309 clusterloader.go:249] Test Finished
I0612 14:54:18.302634   27309 clusterloader.go:250]   Test: ../aws-ebs-csi-driver/hack/e2e/metrics/cl2-config.yaml
I0612 14:54:18.302640   27309 clusterloader.go:251]   Status: Success
I0612 14:54:18.302645   27309 clusterloader.go:255] --------------------------------------------------------------------------------

2023-06-12 14:54:33 [INFO] - Clusterloader2 tests completed successfully
2023-06-12 14:54:33 [INFO] - Checking if S3 bucket aws-ebs-csi-driver-metrics exists
2023-06-12 14:54:34 [INFO] - Configuring bucket policy to allow Lambda access
2023-06-12 14:54:36 [INFO] - Setting lifecycle policy on S3 bucket aws-ebs-csi-driver-metrics
2023-06-12 14:54:37 [INFO] - Configuring bucket notification to trigger Lambda function
2023-06-12 14:54:39 [INFO] - Uploading metrics to S3 bucket aws-ebs-csi-driver-metrics
```